Time: 4 ms (64.77%) | Memory: 77.1 MB (17.39%) - LeetSync

+++ 878-shifting-letters/shifting-letters.cpp
@@ -5,7 +5,7 @@ public:
         vector<long long>pSum(n, 0);
         pSum[n-1] = shifts[n-1];
         for(int i=n-2; i>=0; i--){
-            pSum[i] = (long long)pSum[i+1] + shifts[i];
+            pSum[i] = pSum[i+1] + shifts[i];
         }
 
         for(int i=0; i<n; i++){
